+++ src/main/java/fr/mrcraftcod/gunterdiscord/settings/configs/MegaWarnTimeConfig.java
@@ -17,7 +17,7 @@ public class MegaWarnTimeConfig extends ValueConfiguration
 	{
 		try
 		{
-			super.setValue(guild, Integer.parseInt(value.toString()));
+			super.setValue(guild, Double.parseDouble(value.toString()));
 		}
 		catch(Exception e)
 		{
